requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.133
botocore==1.29.133
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
diagrams==0.23.3
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.5.9
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-api-core==2.11.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.13.1rc0
langchain==0.0.168
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pansi==2020.7.3
Pillow==9.5.0
protobuf==4.23.0
py2neo==2021.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
Pygments==2.15.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.13
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc0
termcolor==2.3.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.1
typed-ast==1.5.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
